### 2020-02-13 #2099 @whedon
[PRE REVIEW]: CloudnetPy: A Python package for processing cloud remote sensing data
**Submitting author:** @tukiains (<a href="http://orcid.org/0000-0002-0651-4622">Simo Tukiainen</a>)
**Repository:** <a href="https://github.com/actris-cloudnet/cloudnetpy" target ="_blank">https://github.com/actris-cloudnet/cloudnetpy</a>
**Version:** v1.0.7
**Reviewers:** josephhardinee
