### 2020-02-18 #2099 @jedbrown
@scivision @josephhardinee :wave: Would you be available to review this submission for JOSS?
