### 2020-02-18 #2099 @josephhardinee
Sure

Sent from my iPhone

> On Feb 18, 2020, at 7:14 AM, Jed Brown <notifications@github.com> wrote:
>
> ﻿
> @scivision @josephhardinee 👋 Would you be available to review this submission for JOSS?
>
