### 2020-02-27 #2123 @jedbrown
@josephhardinee :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the CloudnetPy repository). I'll be watching this thread if you have any questions.
