### 2020-03-03 #2123 @jedbrown
@ekschuman :wave: Welcome to JOSS and thanks for agreeing to review! (I understand you are available and willing, per Joseph.  Please let me know here or by email if that is not accurate or you need any accommodation.)

The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the CloudnetPy repository). I'll be watching this thread if you have any questions.
