### 2020-04-07 7160070
Add keep_uuid option to ceilo2nc
