### 2020-04-07 79d22af
Add keep_uuid to categorize and product API functions
