### 2020-04-11 3cab7ac
Use join to create rpg file paths
