+++ cloudnetpy/version.py
@@ -1,4 +1,4 @@
 MAJOR = 1
 MINOR = 0
-PATCH = 7
+PATCH = 8
 __version__ = '%d.%d.%d' % (MAJOR, MINOR, PATCH)
