### 2020-04-21 e8e2667
Move processing scripts into own repository
