+++ bin/process_all.py
@@ -1,374 +1,0 @@
-"""Testing script for Cloudnet processing.
-
-Creates similar directory structure and file names as the
-Matlab processing environment.
-
-"""
-import os
-import sys
-from pathlib import Path
-sys.path.insert(0, str(Path(__file__).resolve().parents[1]))
-import fnmatch
-import gzip
-import shutil
-import importlib
-import configparser
-import datetime
-from collections import namedtuple
-from cloudnetpy.categorize import generate_categorize
-from cloudnetpy.instruments import mira2nc, rpg2nc, ceilo2nc
-from cloudnetpy.plotting import generate_figure
-from cloudnetpy import utils
-
-config = configparser.ConfigParser()
-config.read('config.ini')
-
-INPUT_ROOT = f"{config['PATH']['input_root_path']}{config['SITE']['dir_name']}/"
-OUTPUT_ROOT = f"{config['PATH']['output_root_path']}{config['SITE']['dir_name']}/"
-
-
-def main():
-    """ Main Cloudnet processing function."""
-
-    _read_site_altitude()
-
-    start_date = _period_to_date('PERIOD_START')
-    end_date = _period_to_date('PERIOD_END')
-
-    for date in utils.date_range(start_date, end_date):
-        dvec = date.strftime("%Y%m%d")
-        print('Date: ', dvec)
-        for processing_type in ('radar', 'lidar', 'categorize'):
-            _run_processing(processing_type, dvec)
-        for product in ('classification', 'iwc-Z-T-method',
-                        'lwc-scaled-adiabatic', 'drizzle'):
-            try:
-                _process_product(product, dvec)
-            except RuntimeError as error:
-                print(error)
-        print(' ')
-
-
-def _period_to_date(section_name):
-    date = [config.getint(section_name, key) for key in ('year', 'month', 'day')]
-    return datetime.date(*date)
-
-
-def _run_processing(process_type, dvec):
-    module = importlib.import_module(__name__)
-    try:
-        getattr(module, f"_process_{process_type}")(dvec)
-    except RuntimeError as error:
-        print(error)
-
-
-def _process_radar(dvec):
-    def _find_uncalibrated_mira_file():
-        input_path = _find_uncalibrated_path(instrument, dvec)
-        try:
-            file = _find_file(input_path, f"*{dvec}*nc")
-        except FileNotFoundError:
-            try:
-                file = _find_file(input_path, f"*{dvec}*.gz")
-            except FileNotFoundError as error:
-                raise error
-            file = gz_to_nc(file)
-        return file
-
-    instrument = 'radar'
-    if not _should_we_process(instrument):
-        return
-    output_file = _build_calibrated_file_name(instrument, dvec)
-    if config['INSTRUMENTS'][instrument] == 'mira':
-        try:
-            input_file = _find_uncalibrated_mira_file()
-        except FileNotFoundError:
-            raise RuntimeError('Abort: Missing uncalibrated mira file.')
-        if _is_good_to_process(instrument, output_file):
-            print(f"Calibrating mira cloud radar..")
-            mira2nc(input_file, output_file, config['SITE'])
-            _discard_uncompressed_radar_file(input_file)
-    elif config['INSTRUMENTS'][instrument] == 'rpg-fmcw-94':
-        rpg_path = _build_uncalibrated_rpg_path(dvec)
-        try:
-            _ = _find_file(rpg_path, '*.LV1')
-        except FileNotFoundError:
-            raise RuntimeError('Abort: Missing uncalibrated rpg .LV1 files.')
[…]
