+++ bin/config.ini
@@ -1,43 +1,0 @@
-[SITE]
-name = Mace Head
-dir_name = mace-head
-
-[INSTRUMENTS]
-radar = mira
-lidar = chm15k
-mwr = hatpro
-model = ecmwf
-
-[PERIOD_START]
-year = 2019
-month = 5
-day = 17
-
-[PERIOD_END]
-year = 2019
-month = 5
-day = 18
-
-[PROCESS_LEVEL]  # 0=no, 1=if missing, 2=yes
-radar = 2
-lidar = 2
-categorize = 2
-classification = 2
-iwc-Z-T-method = 2
-lwc-scaled-adiabatic = 2
-drizzle = 2
-
-[QUICKLOOK_LEVEL] # 0=no, 1=if missing, 2=yes
-categorize = 2
-classification = 2
-iwc-Z-T-method = 2
-lwc-scaled-adiabatic = 2
-drizzle = 2
-
-[MISC]
-show_plot = no
-save_compressed_mira_raw_only = yes
-
-[PATH]
-input_root_path = /home/korpinen/Documents/ACTRIS/cloudnetpy/tests/source_data/
-output_root_path = /home/korpinen/Documents/ACTRIS/cloudnetpy/tests/source_data/
