+++ MANIFEST.in
@@ -1,1 +1,2 @@
 recursive-exclude tests *
+include cloudnetpy/products/mie_lu_tables.nc
