### 2020-04-28 c87111b
Use cloudnet_file_type. This breaks comparison with legacy files.
