### 2020-04-29 c19c866
Add cloud fraction for plots
