### 2020-04-30 e719aaf
Allow slight deviations in rpg header
