+++ tests/data_quality/data_quality_config.ini
@@ -34,8 +34,8 @@ v = -10, 10
 width = 0, 10
 ldr = -40, 20
 beta_raw = -1e-4, 1
-beta = 1e-12, 1
-beta_smooth = 1e-12, 1
+beta = 1e-13, 1
+beta_smooth = 1e-13, 1
 v_sigma = 0, 5
 insect_prob = 0, 1
 is_rain = 0, 1
