### 2020-05-03 5c7962d
Remove unsed code
