### 2020-05-03 a57504b
Limit values instead of raising error
