+++ .github/workflows/test.yml
@@ -1,0 +1,33 @@
+name: CloudnetPy CI
+
+on: [push]
+
+jobs:
+  build:
+
+    timeout-minutes: 5
+
+    runs-on: ubuntu-latest
+
+    strategy:
+      matrix:
+        python-version: [3.6]
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v1
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        pip install --upgrade pip
+        pip install .
+    - name: Lint with pylint
+      run: |
+        pip install pylint
+        pylint **/*.py --errors-only --ignored-modules=netCDF4
+    - name: Test with pytest
+      run: |
+        pip install pytest
+        pytest --cov=cloudnetpy tests/unit/
