+++ setup.py
@@ -21,8 +21,7 @@ setup(
     packages=find_packages(),
     include_package_data=True,
     python_requires='>=3.6',
-    install_requires=['numpy>=1.16', 'scipy>=1.2', 'netCDF4>=1.4.2',
-                      'matplotlib>=3.0.2', 'requests>=2.21'],
+    install_requires=['scipy', 'netCDF4', 'matplotlib', 'requests'],
     classifiers=[
         "Development Status :: 4 - Beta",
         "Programming Language :: Python :: 3.6",
