### 2020-05-04 9bc47cf
Fix pylint call
