+++ .github/workflows/test.yml
@@ -31,3 +31,10 @@ jobs:
       run: |
         pip install pytest pytest-cov
         pytest --cov=cloudnetpy tests/unit/
+    - name: Run e2e-test
+      run: |
+        cd tests/
+        python3 run_example_processing_and_tests.py
+
+
+
