### 2020-05-04 14afa46
Check exit codes
