+++ tests/run_example_processing_and_tests.py
@@ -8,7 +8,7 @@ import importlib
 from cloudnetpy.instruments import mira2nc
 from cloudnetpy.instruments import ceilo2nc
 from cloudnetpy.categorize import generate_categorize
-from tests import check_data_quality, check_metadata, run_unit_tests, utils
+from tests import check_data_quality, check_metadata, utils
 
 PROCESS = True
 
