### 2020-05-11 81424de
Return UUID of generated nc files
