+++ cloudnetpy/instruments/ceilo.py
@@ -44,7 +44,7 @@ def ceilo2nc(input_file, output_file, site_meta, keep_uuid=False):
     _append_data(ceilo, beta_variants)
     _append_height(ceilo, site_meta['altitude'])
     output.update_attributes(ceilo.data, ATTRIBUTES)
-    _save_ceilo(ceilo, output_file, site_meta['name'], keep_uuid)
+    return _save_ceilo(ceilo, output_file, site_meta['name'], keep_uuid)
 
 
 def _initialize_ceilo(file, site_name):
@@ -102,6 +102,7 @@ def _save_ceilo(ceilo, output_file, location, keep_uuid):
     dims = {'time': len(ceilo.time),
             'range': len(ceilo.range)}
     rootgrp = output.init_file(output_file, dims, ceilo.data, keep_uuid)
+    uuid = rootgrp.file_uuid
     output.add_file_type(rootgrp, 'lidar')
     if hasattr(ceilo, 'dataset'):
         output.copy_variables(ceilo.dataset, rootgrp, ('wavelength',))
@@ -111,6 +112,7 @@ def _save_ceilo(ceilo, output_file, location, keep_uuid):
     rootgrp.history = f"{utils.get_time()} - ceilometer file created"
     rootgrp.source = ceilo.model
     rootgrp.close()
+    return uuid
 
 
 ATTRIBUTES = {
