+++ cloudnetpy/products/iwc.py
@@ -48,8 +48,9 @@ def generate_iwc(categorize_file, output_file, keep_uuid=False):
     _append_iwc_sensitivity(iwc_data)
     _append_iwc_status(iwc_data, ice_class)
     output.update_attributes(iwc_data.data, IWC_ATTRIBUTES)
-    output.save_product_file('iwc', iwc_data, output_file, keep_uuid)
+    uuid = output.save_product_file('iwc', iwc_data, output_file, keep_uuid)
     iwc_data.close()
+    return uuid
 
 
 class IwcSource(DataSource):
