+++ cloudnetpy/products/classification.py
@@ -32,8 +32,9 @@ def generate_classification(categorize_file, output_file, keep_uuid=False):
     status = _get_detection_status(categorize_bits)
     data_handler.append_data(status, 'detection_status')
     output.update_attributes(data_handler.data, CLASSIFICATION_ATTRIBUTES)
-    output.save_product_file('classification', data_handler, output_file, keep_uuid)
+    uuid = output.save_product_file('classification', data_handler, output_file, keep_uuid)
     data_handler.close()
+    return uuid
 
 
 def _get_target_classification(categorize_bits):
