+++ cloudnetpy/instruments/mira.py
@@ -40,7 +40,7 @@ def mira2nc(mmclx_file, output_file, site_meta, rebin_data=False, keep_uuid=Fals
     raw_mira.screen_by_snr(snr_gain)
     raw_mira.add_meta()
     output.update_attributes(raw_mira.data, MIRA_ATTRIBUTES)
-    _save_mira(mmclx_file, raw_mira, output_file, keep_uuid)
+    return _save_mira(mmclx_file, raw_mira, output_file, keep_uuid)
 
 
 class Mira(DataSource):
@@ -119,6 +119,7 @@ def _save_mira(mmclx_file, raw_radar, output_file, keep_uuid):
     dims = {'time': len(raw_radar.time),
             'range': len(raw_radar.range)}
     rootgrp = output.init_file(output_file, dims, raw_radar.data, keep_uuid)
+    uuid = rootgrp.file_uuid
     fields_from_raw = ('nfft', 'prf', 'nave', 'zrg', 'rg0', 'drg')
     output.copy_variables(netCDF4.Dataset(mmclx_file), rootgrp, fields_from_raw)
     output.add_file_type(rootgrp, 'radar')
@@ -128,6 +129,7 @@ def _save_mira(mmclx_file, raw_radar, output_file, keep_uuid):
     rootgrp.history = f"{utils.get_time()} - radar file created"
     rootgrp.source = raw_radar.source
     rootgrp.close()
+    return uuid
 
 
 def _find_measurement_date(raw_radar):
