### 2020-05-11 297ad8b
Test that functions return UUIDs
