+++ tests/__init__.py
@@ -1,1 +1,1 @@
-from .api import check_data_quality, check_metadata, run_unit_tests
+from .api import check_data_quality, check_metadata, check_is_valid_uuid, run_unit_tests
