+++ cloudnetpy/instruments/rpg.py
@@ -23,6 +23,9 @@ def rpg2nc(path_to_l1_files, output_file, site_meta, keep_uuid=False):
         keep_uuid (bool, optional): If True, keeps the UUID of the old file,
             if that exists. Default is False when new UUID is generated.
 
+    Returns:
+        uuid: UUID of the generated file.
+
     Raises:
         RuntimeError: Failed to read the binary data.
 
