+++ cloudnetpy/instruments/ceilo.py
@@ -27,6 +27,9 @@ def ceilo2nc(input_file, output_file, site_meta, keep_uuid=False):
             (metres above mean sea level).
         keep_uuid (bool, optional): If True, keeps the UUID of the old file,
             if that exists. Default is False when new UUID is generated.
+    
+    Returns:
+        uuid: UUID of the generated file.
 
     Raises:
         RuntimeError: Failed to read or process raw ceilometer data.
