### 2020-05-11 f0e927d
Fix docstring returns
