+++ cloudnetpy/instruments/mira.py
@@ -26,7 +26,7 @@ def mira2nc(mmclx_file, output_file, site_meta, rebin_data=False, keep_uuid=Fals
             if that exists. Default is False when new UUID is generated.
     
     Returns:
-        uuid: UUID of the generated file.
+        str: UUID of the generated file.
 
     Examples:
           >>> from cloudnetpy.instruments import mira2nc
