+++ cloudnetpy/categorize/categorize.py
@@ -25,7 +25,7 @@ def generate_categorize(input_files, output_file, keep_uuid=False):
             if that exists. Default is False when new UUID is generated.
     
     Returns:
-        uuid: UUID of the generated file.
+        str: UUID of the generated file.
 
     Raises:
         RuntimeError: Failed to create the categorize file.
