+++ cloudnetpy/version.py
@@ -1,4 +1,4 @@
 MAJOR = 1
-MINOR = 0
-PATCH = 9
+MINOR = 1
+PATCH = 0
 __version__ = '%d.%d.%d' % (MAJOR, MINOR, PATCH)
