### 2020-05-11 19532f8
Add publishing to pypi when released
