### 2020-06-02 a19ac7a
Add option to plot without title
