### 2020-06-02 772633e
Add option to plot without title
