### 2020-06-03 actris-cloudnet/cloudnetpy#7 @ekschuman
Quickstart Guide Example File Clarity
Review: openjournals/joss-reviews#2123

I worked through the quickstart guide on the official documentation.  I had some difficulty with using the provided example files to follow the quickstart guide.  For me getting started in the documentation, it wasn't obvious which file belonged to which category when getting to the generate_categorize function (lidar, radar, model, mwr).  Explicitly mentioning the provided example files in the code examples would probably be a smoother learning experience for new users.

For the steps where mira2nc and ceilo2nc are called, it seemed a bit confusing that the functions are named x2nc, but the provided example files are already in the netCDF4 format, with the nc extension.  I was looking for example files with extensions similar to the example code snippets, but I didn't find any. I then assumed that the .nc files were the already-processed versions, and then tried to use them in generate_categorize directly.  I eventually found that you had to use mira2nc and ceilo2nc on the provided .nc files first, then use the outputs in generate_categorize.

Additionally, in the quickstart documentation, it could be helpful to clarify that all 4 files are necessary for generate_categorize.
