### 2020-06-03 actris-cloudnet/cloudnetpy#8 @ekschuman
Statement of Need in Documentation
Review: openjournals/joss-reviews#2123

The paper associated with this project has a clear statement of need ("Do the authors clearly state what problems the software is designed to solve and who the target audience is?"), but there doesn't seem to be as clear a statement of need for the software in the official project documentation.  The documentation mentions briefly that the package will be used in the operational Cloudnet processing, but the associated paper seems to go into greater detail about why this package is needed and the advantages it provides, as well as the potential audience/users for this software.
