### 2020-06-08 56b02c3
Update quickstart doc
