+++ docs/source/quickstart.rst
@@ -9,43 +9,43 @@ always try `these example files <http://devcloudnet.fmi.fi/files/cloudnetpy_test
 Radar processing
 ----------------
 
-In the first example we convert a raw METEK MIRA-36 netCDF file into
+In the first example we convert a raw METEK MIRA-36 cloud radar file into
 Cloudnet netCDF file that can be used in further processing steps.
 
 .. code-block:: python
 
     from cloudnetpy.instruments import mira2nc
-    mira2nc('raw_radar.mmclx', 'radar.nc', {'name': 'Mace-Head'})
+    uuid = mira2nc('raw_mira_radar.mmclx', 'radar.nc', {'name': 'Mace-Head'})
+
+where ``uuid`` is an unique identifier for the generated ``radar.nc`` file.
+For more information, see `API reference <api.html#instruments.mira2nc>`__ for this function.
 
 Lidar processing
 ----------------
 
-Next we convert a raw Vaisala ceilometer text file into netCDF (and process
-the signal-to-noise screened backscatter).
+Next we convert a raw Jenoptik CHM15k ceilometer (lidar) file into Cloudnet netCDF file
+and process the signal-to-noise screened backscatter. Also this converted lidar
+file will be needed later.
 
 .. code-block:: python
 
     from cloudnetpy.instruments import ceilo2nc
-    ceilo2nc('vaisala.txt', 'vaisala.nc', {'name':'Kumpula', 'altitude':53})
+    uuid = ceilo2nc('raw_chm15k_lidar.nc', 'lidar.nc', {'name':'Mace-Head', 'altitude':5})
 
-The same function can handle also Jenoptik CHM15k files.
-
-.. code-block:: python
-
-    ceilo2nc('jenoptik_chm15k.nc', 'jenoptik.nc', {'name':'Mace Head', 'altitude':16})
-
+where ``uuid`` is an unique identifier for the generated ``lidar.nc`` file.
+For more information, see `API reference <api.html#instruments.ceilo2nc>`__ for this function.
 
 MWR processing
 --------------
 
-Processing of multi-channel HATPRO microwave radiometer (MWR) data is not yet part of CloudnetPy.
+Processing of multi-channel HATPRO microwave radiometer (MWR) data is not part of CloudnetPy.
 Thus, site operators need to run custom processing software to retrieve integrated liquid
 water path (LWP) from raw HATPRO measurements.
 
 However, with a 94 GHz RPG cloud radar, a separate MWR instrument is not necessarely
 required. RPG radars contain single MWR channel providing a rough estimate
 of LWP, which can be used in CloudnetPy. Nevertheless, it is always
-recommended to equip measurement site with a dedicated multi-channel
+recommended to equip a measurement site with a dedicated multi-channel
 radiometer if possible.
 
 Model data
@@ -64,8 +64,13 @@ Any model file can be used in the processing but the recommended order is
 Categorize processing
 ---------------------
 
-In the next example we create a categorize file from already
-calibrated measurement files.
+After processing the raw radar and raw lidar files, and acquiring
+the model and mwr files, a Cloudnet categorize file can be created.
+
+In the next example we create a categorize file starting from the
+``radar.nc`` and ``lidar.nc`` files generated above. The required
+``ecmwf_model.nc`` and ``hatpro_mwr.nc`` files are
+included in the provided `example input files <http://devcloudnet.fmi.fi/files/cloudnetpy_test_input_files.zip>`_.
 
 .. code-block:: python
 
@@ -73,23 +78,29 @@ calibrated measurement files.
    input_files = {
        'radar': 'radar.nc',
        'lidar': 'lidar.nc',
[…]
