### 2020-06-08 3b20504
Use changed input file names
