### 2020-06-08 actris-cloudnet/cloudnetpy#8 @tukiains
I agree, this could be elaborated more in the documentation. I modified the overview page and added a statement of need section (https://github.com/actris-cloudnet/cloudnetpy/commit/c9129fe084eeae2b95dbabd4601935d1d915965a).
