### 2020-08-03 #2123 @josephhardinee
I apologize, I was out at the end of last week and this kinda slipped off my radar with things going on. I'll get right back on this as I still have some notes from when I first  started on it. I'll also ping Eddie as I think he was out on vacation last week .
