### 2020-08-10 22fc3d4
Fix colorbar ticks
