+++ cloudnetpy/version.py
@@ -1,4 +1,4 @@
 MAJOR = 1
 MINOR = 2
-PATCH = 1
+PATCH = 2
 __version__ = '%d.%d.%d' % (MAJOR, MINOR, PATCH)
