### 2020-08-10 actris-cloudnet/cloudnetpy#9 @tukiains
This was caused by a bug in the colorbar ticks. Thanks for spotting! It is fixed in https://github.com/actris-cloudnet/cloudnetpy/commit/22fc3d4a24660adcd428096e423bb6e14b368b89, included in https://github.com/actris-cloudnet/cloudnetpy/releases/tag/v1.2.2.
