### 2020-08-20 1d8e093
Improve docstrings
