### 2020-08-24 6341bd7
Remove redundant attribute
