### 2020-08-25 28794fc
Document radar.type attribute better
