### 2020-08-27 b6f6c2f
Remove duplicate note
