+++ cloudnetpy/version.py
@@ -1,4 +1,4 @@
 MAJOR = 1
 MINOR = 2
-PATCH = 2
+PATCH = 3
 __version__ = '%d.%d.%d' % (MAJOR, MINOR, PATCH)
