### 2020-09-02 #2123 @tukiains
The proof looks good to me.
