### 2020-09-02 becc547
Make sure the string conversion works
