+++ cloudnetpy/version.py
@@ -1,4 +1,4 @@
 MAJOR = 1
 MINOR = 2
-PATCH = 3
+PATCH = 4
 __version__ = '%d.%d.%d' % (MAJOR, MINOR, PATCH)
