### 2020-09-02 #2123 @jedbrown
Reads well to me. In order to proceed with acceptance, could you please:
