### 2020-09-02 #2123 @jedbrown
And should the version be v1.0.7 or one of the later releases (v1.2.4 looks to be the most recent)?
