### 2020-09-03 #2123 @tukiains
Yes, the correct version is the latest (v1.2.4). This tagged release is archived in Zenodo with the correct metadata.

The DOI of the archived version is [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4011843.svg)](https://doi.org/10.5281/zenodo.4011843)
