### 2020-09-07 #2123 @tukiains
Thank you @jedbrown and @kthyng for editing and publishing the paper, and big thanks @josephhardinee and @ekschuman for reviewing!
